move worker unenv files
